+++ apt/apt-pkg/srcrecords.h
@@ -32,7 +32,7 @@ class pkgSrcRecords
    struct File
    {
       string MD5Hash;
-      unsigned long Size;
+      unsigned long long Size;
       string Path;
       string Type;
    };
@@ -61,8 +61,8 @@ class pkgSrcRecords
       
       virtual bool Restart() = 0;
       virtual bool Step() = 0;
-      virtual bool Jump(unsigned long Off) = 0;
-      virtual unsigned long Offset() = 0;
+      virtual bool Jump(unsigned long long Off) = 0;
+      virtual unsigned long long Offset() = 0;
       virtual string AsStr() = 0;
       
       virtual string FileName() const = 0;
